Cleaned up casting of param primitive cruft from ctor cleanup

+++ MatchMakerTests/Data Bags/PlayerTests.cs
@@ -10,7 +10,7 @@ namespace MatchMakerTests.Data_Bags
         [TestMethod]
         public void ShouldEquatePlayerCorrectly()
         {
-            Player player = new Player((int) 2, (double) 50d, (int) 499);
+            Player player = new Player(2, 50d, 499);
 
             bool value = player.Equals(player);
 
@@ -20,8 +20,8 @@ namespace MatchMakerTests.Data_Bags
         [TestMethod]
         public void ShouldFindNotEqualsToo()
         {
-            Player player = new Player((int) 2, (double) 50d, (int) 499);
-            Player otherPlayer = new Player((int) 3, (double) 50d, (int) 499);
+            Player player = new Player(2, 50d, 499);
+            Player otherPlayer = new Player(3, 50d, 499);
 
             bool value = otherPlayer.Equals(player);
 
@@ -31,7 +31,7 @@ namespace MatchMakerTests.Data_Bags
         [TestMethod]
         public void ShouldNotEquateNull()
         {
-            Player player = new Player((int) 2, (double) 50d, (int) 499);
+            Player player = new Player(2, 50d, 499);
 
             bool value = player.Equals(null);
 
@@ -41,9 +41,9 @@ namespace MatchMakerTests.Data_Bags
         [TestMethod]
         public void ShouldEqualOnType()
         {
-            Player player = new Player((int) 2, (double) 50d, (int) 499);
+            Player player = new Player(2, 50d, 499);
 
-            player.Equals(new Player((int) 4, (double) 50d, (int) 499)).Should().BeFalse();
+            player.Equals(new Player(4, 50d, 499)).Should().BeFalse();
         }
 
         // 60+  - high
@@ -56,10 +56,10 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnTrueForSameWinRateCategory5()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 65.0d, (int) 499);
+            Player player1 = new Player(1, 65.0d, 499);
 
             // act
-            bool result = player1.IsSameWinRateCategory(new Player((int) 2, (double) 75d, (int) 499));
+            bool result = player1.IsSameWinRateCategory(new Player(2, 75d, 499));
 
             // assert
             result.Should().BeTrue();
@@ -69,10 +69,10 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnTrueForSameWinRateCategory4()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 55.0d, (int) 499);
+            Player player1 = new Player(1, 55.0d, 499);
 
             // act
-            bool result = player1.IsSameWinRateCategory(new Player((int) 4, (double) 59d, (int) 499));
+            bool result = player1.IsSameWinRateCategory(new Player(4, 59d, 499));
 
             // assert
             result.Should().BeTrue();
@@ -82,10 +82,10 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnTrueForSameWinRateCategory3()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 50.0d, (int) 499);
+            Player player1 = new Player(1, 50.0d, 499);
 
             // act
-            bool result = player1.IsSameWinRateCategory(new Player((int) 5, (double) 54.9d, (int) 499));
+            bool result = player1.IsSameWinRateCategory(new Player(5, 54.9d, 499));
 
             // assert
             result.Should().BeTrue();
@@ -95,10 +95,10 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnTrueForSameWinRateCategory2()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 45.0d, (int) 499);
+            Player player1 = new Player(1, 45.0d, 499);
 
             // act
-            bool result = player1.IsSameWinRateCategory(new Player((int) 5, (double) 49.999d, (int) 499));
+            bool result = player1.IsSameWinRateCategory(new Player(5, 49.999d, 499));
 
             // assert
             result.Should().BeTrue();
@@ -108,10 +108,10 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnTrueForSameWinRateCategory1()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 40, (int) 499);
+            Player player1 = new Player( 1, 40,  499);
 
             // act
-            bool result = player1.IsSameWinRateCategory(new Player((int) 3, (double) 44.999d, (int) 499));
+            bool result = player1.IsSameWinRateCategory(new Player(3, 44.999d, 499));
 
             // assert
             result.Should().BeTrue();
@@ -199,10 +199,10 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnFalseForNotSameWinRateCategory()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 45, (int) 499);
+            Player player1 = new Player(1, 45, 499);
 
             // act
-            bool result = player1.IsSameWinRateCategory(new Player((int) 3, (double) 23.3d, (int) 499));
+            bool result = player1.IsSameWinRateCategory(new Player(3, 23.3d, 499));
             result.Should().BeFalse();
         }
 
@@ -210,8 +210,8 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnTrueForSameWinRateCategoryFromPlayer()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 40, (int) 499);
-            Player player2 = new Player((int) 1, (double) 43, (int) 499);
+            Player player1 = new Player(1, 40, 499);
+            Player player2 = new Player(1, 43, 499);
 
             // act
             bool result = player1.IsSameWinRateCategory(player2);
@@ -224,8 +224,8 @@ namespace MatchMakerTests.Data_Bags
         public void ShouldReturnFalseForNotSameWinRateCategoryFromPlayer()
         {
             // arrange
-            Player player1 = new Player((int) 1, (double) 45, (int) 499);
-            Player player2 = new Player((int) 2, (double) 20.0d, (int) 499);
+            Player player1 = new Player(1, 45, 499);
+            Player player2 = new Player(2, 20.0d, 499);
 
             // act
             bool result = player1.IsSameWinRateCategory(player2);
